build: gyp modifications for v4.2.5
Add new source files and force the library to be built as static.
The `<library)` was working previously because the library was only
build from `nsolid-node` only and it inherited the `static_library` from
node itself. As now it's used from the bindings for the
`scarab-zmq-agent` addon, and nodejs addons are built as shared
libraries, it needs to be forced.

PR-URL: https://github.com/nodesource/libzmq/pull/4
Reviewed-by: Trevor Norris <trev.norris@gmail.com>